Restore separate Perf/Stopwatch impls on Windows
