+++ src/frontend/dune
@@ -2,4 +2,4 @@
  (name frontend)
  (wrapped false)
  (preprocess (pps ppx_parser))
- (libraries camlp-streams perf))
+ (libraries camlp-streams perf stopwatch))
