+++ src/perf/perf.unix.ml
@@ -1,0 +1,1 @@
+let time = Perf_unix.time
