+++ src/perf/perf.win.ml
@@ -1,0 +1,1 @@
+let time = Perf_win.time
