+++ src/perf/unix/dune
@@ -1,0 +1,4 @@
+(library
+ (name perf_unix)
+ (enabled_if
+  (= %{os_type} Unix)))
